Clear the clipboard on startup.
Since there is no event API for clipboard events, we resort on
polling to fetch clipboard changes. This polling process will
store the last seen hashed value and when it changes, it will
share the content. The problem is that if the user starts the
process, for instance, with "hello" on the clipboard, and then
copy "hello" again, this value will not be sent because it is
the same as the "previous" one.